feat: Change weight input format from decimal to integer
This commit modifies the input format for weight measurements in the Paradigm Scale application from decimal to integer. The decision to eliminate decimal points aligns with conventional practices among farmers, who typically focus on whole numbers during the harvest process.

By using integer values, the app enables quicker and more straightforward calculations, particularly in field environments where fast decision-making is crucial. This change enhances user experience by providing a more practical approach to weight measurement that reflects standard agricultural practices.

+++ app/src/main/java/com/example/weighingscale/data/model/BatchDetail.java
@@ -19,7 +19,7 @@ public class BatchDetail {
     public String id;
     public String batch_id;
     public Date datetime;
-    public Double amount;
+    public int amount;
 
     // Constructor
     public BatchDetail() {
@@ -35,7 +35,7 @@ public class BatchDetail {
         return datetime;
     }
 
-    public Double getAmount() {
+    public int getAmount() {
         return amount;
     }
 }
